시스템관리 - masterDetail row 하나만 열리도록 수정
